requirements.txt
jinja2==2.11.2
pytz==2019.3
slackclient==2.5.0
